Play.js file added
The js in the play.js file has been taken from a previous version of this project that has now been aborted.
A delay in receiving feedback from my first project meant that I had already started my second milestone project.
The feedback has indicated to me that there are ways I can improve my working practice and based on this I took the decision to start the project again.
While much of the html and css is being rewritten, the js was working well so I have chosen to keep this as it is for now.

However, this does mean that there is a lack of recent commits for this section of the project.